requirements.txt
google-generativeai>=0.6.15
python-dotenv>=1.0.0
colorama>=0.4.6


